Validar Admin delete Users

+++ Backend/src/services/Users/DeleteUsers_service.ts
@@ -2,6 +2,7 @@ import { PrismaClient } from "@prisma/client";
 import { ExistsError } from "../../error/ExistsError";
 import { Iduser } from "../../types/Task_types";
 import { IdUsuario } from "../../validators/Task/TaskIdValidator";
+import { ExistsMensage } from "../../error/ExistsMensage";
 
 class DeleteUsersService {
   private prisma: PrismaClient;
@@ -13,6 +14,37 @@ class DeleteUsersService {
   async execute(userData: Iduser) {
     // Valida o ID do usuário
     IdUsuario(userData);
+
+    const idExistUser = await this.prisma.user.findFirst({
+      where: {
+        id: userData.id
+      }
+    })
+
+    if (!idExistUser) {
+      throw new ExistsError("ID do usuário não existe no banco de dados");
+
+    }
+
+    const UserAdminOn = await this.prisma.user.findFirst({
+      where: {
+        id: userData.userAdminID 
+      }
+    });
+    
+    if (UserAdminOn && !UserAdminOn.adminUser) {
+      throw new ExistsMensage('Usuário não tem acesso para exluir, somente administradores');
+    }
+
+    const UserxUxers = await this.prisma.user.findFirst({
+      where: {
+        id: userData.id 
+      }
+    });
+    
+    if (UserxUxers?.id === UserAdminOn?.id) {
+      throw new ExistsMensage('Usuário não tem como excluir ele mesmo');
+    }
 
     // Exclua todas as categorias associadas ao usuário
     await this.prisma.category.deleteMany({
@@ -35,17 +67,13 @@ class DeleteUsersService {
       },
     });
 
-
-    const idExistUser = await this.prisma.user.findFirst({
+    // Exclua todas as Fotos associadas ao usuário
+    await this.prisma.imageStorage.deleteMany({
       where: {
-        id: userData.id
-      }
-    })
+        id: userData.id,
+      },
+    });
 
-    if (!idExistUser) {
-      throw new ExistsError("ID do usuário não existe no banco de dados");
-
-    }
     // Exclua o usuário
      await this.prisma.user.delete({
       where: {
